# src/main/resources/openapi/inventario-api.yaml
openapi: 3.0.3
info:
  title: API Sistema de Inventario Multi-Sucursal
  description: >
    Contrato OpenAPI base del backend de inventario multi-sucursal.
    Define endpoints por modulo antes de implementar controllers para mantener
    separacion de responsabilidades y buenas practicas REST.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Servidor local (dev)
  - url: http://backend:8080
    description: Servidor en docker
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operacion exitosa
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time
          example: "2026-04-25T10:30:00Z"
...
    LineaOrdenRequest:
      type: object
      required: [idProducto, cantidadPedida, precioUnitario]
      properties:
        idProducto:
          type: integer
          format: int64
        cantidadPedida:
          type: number
          format: double
          minimum: 0.0001
        precioUnitario:
          type: number
          format: double
          minimum: 0.0001
        descuento:
          type: number
          format: double
          minimum: 0
          maximum: 100
    OrdenCompraRequest:
      type: object
      required: [idProveedor, idSucursal, lineas]
      properties:
        idProveedor:
          type: integer
          format: int64
        idSucursal:
          type: integer
          format: int64
        fechaEstimadaEntrega:
          type: string
          format: date
          nullable: true
        plazoPagoDias:
          type: integer
          minimum: 1
          nullable: true
        lineas:
          type: array
          items:
            $ref: '#/components/schemas/LineaOrdenRequest'
    LineaRecepcionRequest:
      type: object
      required: [idDetalle, cantidadRecibida]
      properties:
        idDetalle:
          type: integer
          format: int64
        cantidadRecibida:
          type: number
          format: double
          minimum: 0
    RecepcionCompraRequest:
      type: object
      required: [idOrden, lineas]
      properties:
        idOrden:
          type: integer
          format: int64
        lineas:
          type: array
          items:
            $ref: '#/components/schemas/LineaRecepcionRequest'
...
paths:
...
  /api/v1/compras:
    get:
      tags: [Compras]
      summary: Listar ordenes de compra
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: sucursalId
          schema:
            type: integer
            format: int64
        - in: query
          name: proveedorId
          schema:
            type: integer
            format: int64
        - in: query
          name: estado
          schema:
            type: string
            enum: [PENDIENTE, RECIBIDA, RECIBIDA_CON_FALTANTES, CANCELADA]
      responses:
        "200":
          description: Ordenes encontradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Compras]
      summary: Crear orden de compra
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrdenCompraRequest'
      responses:
        "201":
          description: Orden creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/compras/{id}:
    get:
      tags: [Compras]
      summary: Obtener orden de compra por ID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Orden encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/compras/{id}/cancelar:
    patch:
      tags: [Compras]
      summary: Cancelar orden pendiente
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Orden cancelada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/compras/{id}/recepcion:
    post:
      tags: [Compras]
      summary: Confirmar recepcion de compra
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecepcionCompraRequest'
      responses:
        "200":
          description: Recepcion confirmada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...